# api/openapi.yaml
openapi: 3.1.0
info:
  title: Civitas API
  version: 0.2.0
  description: Runtime API for the Civitas civic investigation OS.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
  /readyz:
    get:
      operationId: getReady
      responses:
        "200":
          description: Ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
  /api/v1/version:
    get:
      operationId: getVersion
      responses:
        "200":
          description: Version details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionInfo"
  /api/v1/processors:
    get:
      operationId: listProcessors
      responses:
        "200":
          description: Processor registry.
          content:
            application/json:
              schema:
                type: object
                required: [processors]
                properties:
                  processors:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProcessorTool"
  /api/v1/cases:
    get:
      operationId: listCases
      responses:
        "200":
          description: Case list.
          content:
            application/json:
              schema:
                type: object
                required: [cases]
                properties:
                  cases:
                    type: array
                    items:
                      $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Created case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /api/v1/case-states/import:
    post:
      operationId: importCaseState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseState"
      responses:
        "201":
          description: Imported case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /api/v1/cases/{case_id}/documents:
    parameters:
      - $ref: "#/components/parameters/CaseID"
    get:
      operationId: listDocuments
      responses:
        "200":
          description: Documents in a case.
          content:
            application/json:
              schema:
                type: object
                required: [documents]
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Uploaded and processed document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /api/v1/cases/{case_id}/search:
    parameters:
      - $ref: "#/components/parameters/CaseID"
    get:
      operationId: searchCase
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results.
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchResult"
  /api/v1/cases/{case_id}/graph:
    parameters:
      - $ref: "#/components/parameters/CaseID"
    get:
      operationId: getGraph
      responses:
        "200":
          description: Evidence graph.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Graph"
  /api/v1/cases/{case_id}/timeline:
    parameters:
      - $ref: "#/components/parameters/CaseID"
    get:
      operationId: getTimeline
      responses:
        "200":
          description: Timeline events.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/TimelineEvent"
  /api/v1/cases/{case_id}/debug:
    parameters:
      - $ref: "#/components/parameters/CaseID"
    get:
      operationId: debugCase
      responses:
        "200":
          description: Debug document inference state.
          content:
            application/json:
              schema:
                type: object
                required: [documents, count]
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
                  count:
                    type: integer
  /api/v1/cases/{case_id}/state:
    parameters:
      - $ref: "#/components/parameters/CaseID"
    get:
      operationId: getCaseState
      responses:
        "200":
          description: Downloadable case state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseState"
  /api/v1/cases/{case_id}/exports:
    parameters:
      - $ref: "#/components/parameters/CaseID"
    post:
      operationId: createExport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                format:
                  type: string
                  default: markdown
      responses:
        "201":
          description: Generated export.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
  /api/v1/cases/{case_id}/exports/{export_id}:
    parameters:
      - $ref: "#/components/parameters/CaseID"
      - name: export_id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getExport
      responses:
        "200":
          description: Generated export body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
components:
  parameters:
    CaseID:
      name: case_id
      in: path
      required: true
      schema:
        type: string
  schemas:
    Status:
      type: object
      required: [status]
      properties:
        status:
          type: string
    VersionInfo:
      type: object
      required: [version, commit, mode]
      properties:
        version:
          type: string
        commit:
          type: string
        mode:
          type: string
    Case:
      type: object
      required: [id, title, description, created_at, updated_at, document_ids]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        document_ids:
          type: array
          items:
            type: string
    Document:
      type: object
      required:
        - id
        - case_id
        - filename
        - content_type
        - size
        - sha256
        - status
        - shape
        - state
        - confidence
        - fields
        - anomalies
        - entities
        - timeline
        - processors
        - provenance
        - parse
        - created_at
      properties:
        id:
          type: string
        case_id:
          type: string
        filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
          format: int64
        sha256:
          type: string
        status:
          type: string
        shape:
          type: string
        state:
          type: string
        confidence:
          type: number
        text:
          type: string
        preview:
          type: string
        summary:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldInference"
        anomalies:
          type: array
          items:
            $ref: "#/components/schemas/Anomaly"
        entities:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        processors:
          type: array
          items:
            $ref: "#/components/schemas/ProcessorStatus"
        provenance:
          $ref: "#/components/schemas/Provenance"
        parse:
          $ref: "#/components/schemas/ParseMetrics"
        created_at:
          type: string
          format: date-time
    FieldInference:
      type: object
      required: [id, name, type, value, confidence, reason]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        value:
          type: string
        normalized:
          type: string
        confidence:
          type: number
        reason:
          type: string
    Anomaly:
      type: object
      required: [id, code, severity, message, why, next_step, confidence]
      properties:
        id:
          type: string
        code:
          type: string
        severity:
          type: string
        message:
          type: string
        why:
          type: string
        next_step:
          type: string
        confidence:
          type: number
    Provenance:
      type: object
      required:
        [
          schema_version,
          app_version,
          source_id,
          source_name,
          source_sha256,
          parameters,
        ]
      properties:
        schema_version:
          type: string
        app_version:
          type: string
        source_id:
          type: string
        source_name:
          type: string
        source_sha256:
          type: string
        parameters:
          type: array
          items:
            type: string
    ParseMetrics:
      type: object
      required: [duration_ms, size_bucket]
      properties:
        duration_ms:
          type: integer
          format: int64
        size_bucket:
          type: string
        row_count:
          type: integer
        field_count:
          type: integer
    Entity:
      type: object
      required: [id, type, value, confidence, source]
      properties:
        id:
          type: string
        type:
          type: string
        value:
          type: string
        confidence:
          type: number
        source:
          type: string
    TimelineEvent:
      type: object
      required: [id, document_id, when, label, confidence]
      properties:
        id:
          type: string
        document_id:
          type: string
        when:
          type: string
          format: date-time
        label:
          type: string
        confidence:
          type: number
    ProcessorStatus:
      type: object
      required: [name, kind, available, status]
      properties:
        name:
          type: string
        kind:
          type: string
        available:
          type: boolean
        status:
          type: string
        message:
          type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
    ProcessorTool:
      type: object
      required: [name, kind, command, available, description]
      properties:
        name:
          type: string
        kind:
          type: string
        command:
          type: string
        available:
          type: boolean
        description:
          type: string
    SearchResult:
      type: object
      required: [document_id, filename, snippet, score]
      properties:
        document_id:
          type: string
        filename:
          type: string
        snippet:
          type: string
        score:
          type: number
    Graph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
    GraphNode:
      type: object
      required: [id, type, label]
      properties:
        id:
          type: string
        type:
          type: string
        label:
          type: string
    GraphEdge:
      type: object
      required: [id, source_id, target_id, type, weight]
      properties:
        id:
          type: string
        source_id:
          type: string
        target_id:
          type: string
        type:
          type: string
        weight:
          type: number
    Export:
      type: object
      required: [id, case_id, format, path, created_at]
      properties:
        id:
          type: string
        case_id:
          type: string
        format:
          type: string
        path:
          type: string
        body:
          type: string
        created_at:
          type: string
          format: date-time
    CaseState:
      type: object
      required: [schema_version, app_version, exported_at, case, documents]
      properties:
        schema_version:
          type: string
        app_version:
          type: string
        exported_at:
          type: string
          format: date-time
        case:
          $ref: "#/components/schemas/Case"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/StateDocument"
    StateDocument:
      type: object
      required: [document, content_base64, content_sha256]
      properties:
        document:
          $ref: "#/components/schemas/Document"
        content_base64:
          type: string
        content_sha256:
          type: string